Moved DefaultLogger from interpreter-lib to execution
