+++ libs/execution/src/lib/debugging/debug-log-visitor.ts
@@ -4,7 +4,7 @@
 
 import { internalValueToString } from '@jvalue/jayvee-language-server';
 
-import { Logger } from '../logger';
+import { Logger } from '../logging/logger';
 import { Workbook } from '../types';
 import { FileSystem } from '../types/io-types/filesystem';
 import { BinaryFile } from '../types/io-types/filesystem-node-file-binary';
