+++ libs/execution/src/lib/logging/index.ts
@@ -2,4 +2,5 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-export * from './logger-factory';
+export * from './logger';
+export * from './default-logger';
